[codeshuffle] - moved MessagePrinter.h/.cpp
