+++ .gitignore
@@ -551,6 +551,9 @@ lib/cpluff/stamp-h1
 # /xbmc/main/
 /xbmc/main/Makefile
 
+# /xbmc/platform
+/xbmc/platform/Makefile
+
 # /xbmc/messaging/
 /xbmc/messaging/Makefile
 /xbmc/messaging/helpers/Makefile
@@ -569,9 +572,6 @@ lib/cpluff/stamp-h1
 # /xbmc/peripherals/
 /xbmc/peripherals/bus/Makefile
 /xbmc/peripherals/devices/Makefile
-
-# /xbmc/main/posix/
-xbmc/main/posix/Makefile
 
 #/xbmc/utils
 /xbmc/utils/Makefile
