+++ project/VS2010Express/XBMC.vcxproj
@@ -650,7 +650,7 @@
     <ClCompile Include="..\..\xbmc\main\main.cpp">
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Debug Testsuite|Win32'">true</ExcludedFromBuild>
     </ClCompile>
-    <ClCompile Include="..\..\xbmc\main\win32\MessagePrinter.cpp" />
+    <ClCompile Include="..\..\xbmc\platform\win32\MessagePrinter.cpp" />
     <ClCompile Include="..\..\xbmc\main\win32\WinMain.cpp">
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Debug Testsuite|Win32'">true</ExcludedFromBuild>
     </ClCompile>
@@ -1101,7 +1101,7 @@
     <ClInclude Include="..\..\xbmc\interfaces\python\PyContext.h" />
     <ClInclude Include="..\..\xbmc\interfaces\python\PythonInvoker.h" />
     <ClInclude Include="..\..\xbmc\interfaces\python\pythreadstate.h" />
-    <ClInclude Include="..\..\xbmc\main\win32\MessagePrinter.h" />
+    <ClInclude Include="..\..\xbmc\platform\MessagePrinter.h" />
     <ClInclude Include="..\..\xbmc\media\MediaType.h" />
     <ClInclude Include="..\..\xbmc\messaging\ApplicationMessenger.h" />
     <ClInclude Include="..\..\xbmc\messaging\helpers\DialogHelper.h" />
