+++ project/VS2010Express/XBMC.vcxproj.filters
@@ -85,6 +85,12 @@
     <Filter Include="playlists">
       <UniqueIdentifier>{084b5eb0-c9c9-444b-a2ee-dcb4f9537d35}</UniqueIdentifier>
     </Filter>
+    <Filter Include="platform">
+      <UniqueIdentifier>{90dd91e6-b726-4f88-803a-84156a663713}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="platform\win32">
+      <UniqueIdentifier>{9312ce72-49ff-422a-8e3a-05c0a2a534d5}</UniqueIdentifier>
+    </Filter>
     <Filter Include="powermanagement">
       <UniqueIdentifier>{b203de13-ce8c-4b43-9a87-a05e68fbad13}</UniqueIdentifier>
     </Filter>
@@ -3066,8 +3072,8 @@
     <ClCompile Include="..\..\xbmc\main\win32\WinMain.cpp">
       <Filter>main\win32</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\xbmc\main\win32\MessagePrinter.cpp">
-      <Filter>main\win32</Filter>
+    <ClCompile Include="..\..\xbmc\platform\win32\MessagePrinter.cpp">
+      <Filter>platform\win32</Filter>
     </ClCompile>
     <ClCompile Include="..\..\xbmc\messaging\ApplicationMessenger.cpp">
       <Filter>messaging</Filter>
@@ -6067,8 +6073,8 @@
     <ClInclude Include="..\..\xbmc\pictures\PictureScalingAlgorithm.h">
       <Filter>pictures</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\xbmc\main\win32\MessagePrinter.h">
-      <Filter>main\win32</Filter>
+    <ClInclude Include="..\..\xbmc\platform\MessagePrinter.h">
+      <Filter>platform</Filter>
     </ClInclude>
     <ClInclude Include="..\..\xbmc\messaging\ApplicationMessenger.h">
       <Filter>messaging</Filter>
