--- xbmc/main/posix/Makefile.in
@@ -1,7 +1,0 @@
-.SUFFIXES : .cpp 
- 
-SRCS = MessagePrinter.cpp 
-LIB=MessagePrinter.a 
- 
-include @abs_top_srcdir@/Makefile.include 
--include $(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(SRCS)))
