+++ xbmc/platform/Makefile.in
@@ -1,0 +1,8 @@
+.SUFFIXES : .m .mm .cpp
+
+SRCS = posix/MessagePrinter.cpp
+
+LIB=platform.a
+
+include @abs_top_srcdir@/Makefile.include
+-include $(patsubst %.mm,%.P,$(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(SRCS))))
