+++ xbmc/platform/MessagePrinter.h
